+++ dom/test/browser/isolation.ts
@@ -438,10 +438,11 @@ describe('isolation', function() {
       };
     }
 
-    const {sinks, sources, run} = setup(app, {
+    const drivers = {
       DOM: makeDOMDriver(createRenderTarget()),
-      island: sink => {},
-    });
+      island(sink: Stream<Array<Element>>) {},
+    };
+    const {sinks, sources, run} = setup(app, drivers);
 
     sinks.island
       .drop(1)
@@ -669,10 +670,11 @@ describe('isolation', function() {
       };
     }
 
-    const {sinks, sources, run} = setup(IsolatedApp, {
+    const drivers = {
       DOM: makeDOMDriver(createRenderTarget()),
-      triangleElement: sink => {},
-    });
+      triangleElement: (sink: any) => {},
+    };
+    const {sinks, sources, run} = setup(IsolatedApp, drivers);
 
     // Make assertions
     sinks.triangleElement
@@ -1137,7 +1139,7 @@ describe('isolation', function() {
         .startWith(0)
         .take(4);
       const innerDOM$ = xs
-        .combine(x$, child.DOM)
+        .combine<number, VNode>(x$, child.DOM)
         .map(([x, childVDOM]) =>
           div(`.parent${x}`, {key: `key${x}`}, [childVDOM, `${x}`])
         );
@@ -1199,7 +1201,7 @@ describe('isolation', function() {
         .startWith(0)
         .take(4);
       const innerDOM$ = xs
-        .combine(x$, child.DOM)
+        .combine<number, VNode>(x$, child.DOM)
         .map(([x, childVDOM]) =>
           div(`.parent${x}`, {key: `key${x}`}, [childVDOM, `${x}`])
         );
@@ -1531,6 +1533,7 @@ describe('isolation', function() {
               return div([]);
             } else {
               done('This case must not happen.');
+              return div();
             }
           });
         return {DOM: vdom$};